Moved erd test function to test_file_writer and deleted test_plugin.py/CI file
