--- .github/workflows/test_plugin.yml
@@ -1,36 +1,0 @@
-name: test_plugin.py test
-
-on:
-  push:
-    branches:
-    - main
-  pull_request:
-    branches:
-    - main
-
-
-jobs:
-  linux:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ['3.11']
-
-    steps:
-    - uses: actions/checkout@v4
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v5
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install -r requirements.txt
-        python -m pip install opencv-python
-        pip install .
-        pip install graphviz
-        sudo apt-get install graphviz
-    - name: Test reader
-      run: |
-        pip install pytest
-        pytest dsi/tests/test_plugin.py
